# docs/api/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Workipedia API",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
    "/admin/departments": {
      "post": {
        "summary": "부서 등록",
        "deprecated": false,
        "description": "",
        "operationId": "getDepartments",
        "tags": [
          "departments"
        ],
        "parameters": [
          {
            "name": "Content-Type",
            "in": "header",
            "description": "",
            "required": false,
            "example": "application/json",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {}
              },
              "example": {
                "departmentName": "개발 2팀"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DepartmentResponse"
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "summary": "관리자 부서 목록 조회",
        "deprecated": false,
        "description": "",
        "operationId": "getDepartments",
        "tags": [
          "departments"
        ],
        "parameters": [
          {
            "name": "Content-Type",
            "in": "header",
            "description": "",
            "required": false,
            "example": "application/json",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DepartmentResponse"
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/admin/departments/routing-prompt/instruction": {
      "patch": {
        "summary": "역할 배정 공용 프롬프트 수정",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "Content-Type",
            "in": "header",
            "description": "",
            "required": false,
            "example": "application/json",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {}
              },
              "example": {
                "instruction": "개발 1팀은 ERP랑 IT 담당이고 개발 2팀은 QLOLA 담당이야"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                },
                "example": [
                  {
                    "departmentId": 1,
                    "departmentName": "인사 관리팀",
                    "routingPrompt": "인사팀은 인사관리담당"
                  },
                  {
                    "departmentId": 2,
                    "departmentName": "총무1팀",
                    "routingPrompt": "사내 PC 장애, 계정 권한, 네트워크 문제, 업무 시스템 오류 문의를 담당합니다."
                  },
                  {
                    "departmentId": 3,
                    "departmentName": "IT지원팀",
                    "routingPrompt": null
                  },
                  {
                    "departmentId": 4,
                    "departmentName": "개발팀",
                    "routingPrompt": null
                  },
                  {
                    "departmentId": 5,
                    "departmentName": "재무관리팀",
                    "routingPrompt": null
                  },
                  {
                    "departmentId": 6,
                    "departmentName": "개발 1팀",
                    "routingPrompt": "개발 1팀은 ERP랑 IT 담당이고\r\n개발 1팀은 ERD담당\r\n개발 1팀은 ERP랑 IT 담당이고 개발 2팀은 QLOLA 담당이야\r\n개발 1팀은 ERP랑 IT 담당이고 개발 2팀은 QLOLA 담당이야"
                  }
                ]
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "DepartmentResponse": {
        "type": "object",
        "properties": {
          "departmentId": {
            "type": "integer",
            "format": "int64"
          },
          "departmentName": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "servers": []
}